// frontend/src/api/models/LocationResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LocationResponse
 */
export interface LocationResponse {
    /**
     * 
     * @type {number}
     * @memberof LocationResponse
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof LocationResponse
     */
    x: number;
    /**
     * 
     * @type {number}
     * @memberof LocationResponse
     */
    y: number;
    /**
     * 
     * @type {number}
     * @memberof LocationResponse
     */
    z: number;
    /**
     * 
     * @type {string}
     * @memberof LocationResponse
     */
    name: string;
}

/**
 * Check if a given object implements the LocationResponse interface.
 */
export function instanceOfLocationResponse(value: object): value is LocationResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('x' in value) || value['x'] === undefined) return false;
    if (!('y' in value) || value['y'] === undefined) return false;
    if (!('z' in value) || value['z'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function LocationResponseFromJSON(json: any): LocationResponse {
    return LocationResponseFromJSONTyped(json, false);
}

export function LocationResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'x': json['x'],
        'y': json['y'],
        'z': json['z'],
        'name': json['name'],
    };
}

export function LocationResponseToJSON(json: any): LocationResponse {
    return LocationResponseToJSONTyped(json, false);
}

export function LocationResponseToJSONTyped(value?: LocationResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'x': value['x'],
        'y': value['y'],
        'z': value['z'],
        'name': value['name'],
    };
}